Move the classes to other package
