+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/Session.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.sessionid.model;
+package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
