+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/exception/ServerUrlReadingException.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.sessionid.documentreader.exception;
+package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.exception;
 
 public class ServerUrlReadingException extends RuntimeException {
     public ServerUrlReadingException(String message) {
