+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/exception/SessionIdReadingException.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.sessionid.documentreader.exception;
+package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.exception;
 
 public class SessionIdReadingException extends RuntimeException {
     public SessionIdReadingException(String message) {
